Refactor feature flags to a single file
It's a pretty simple implementation right now, no need for the messy
feature/feature style import.
