+++ background/services/chain/index.ts
@@ -23,7 +23,7 @@ import { AssetTransfer } from "../../assets"
 import { HOUR } from "../../constants"
 import { ETH } from "../../constants/currencies"
 import { ETHEREUM, ARBITRUM_ONE, OPTIMISM } from "../../constants/networks"
-import { MULTI_NETWORK as USE_MULTI_NETWORK } from "../../features/features"
+import { MULTI_NETWORK as USE_MULTI_NETWORK } from "../../features"
 import PreferenceService from "../preferences"
 import { ServiceCreatorFunction, ServiceLifecycleEvents } from "../types"
 import { getOrCreateDB, ChainDatabase } from "./db"
