+++ background/services/signing/index.ts
@@ -11,7 +11,7 @@ import { EIP712TypedData, HexString } from "../../types"
 import BaseService from "../base"
 import { ServiceCreatorFunction, ServiceLifecycleEvents } from "../types"
 import ChainService from "../chain"
-import { USE_MAINNET_FORK } from "../../features/features"
+import { USE_MAINNET_FORK } from "../../features"
 import { FORK } from "../../constants"
 import { SigningMethod } from "../../utils/signing"
