+++ ui/pages/Onboarding/OnboardingImportMetamask.tsx
@@ -3,7 +3,7 @@ import { importKeyring } from "@tallyho/tally-background/redux-slices/keyrings"
 import { useHistory } from "react-router-dom"
 import { isValidMnemonic } from "@ethersproject/hdnode"
 import classNames from "classnames"
-import { HIDE_IMPORT_DERIVATION_PATH } from "@tallyho/tally-background/features/features"
+import { HIDE_IMPORT_DERIVATION_PATH } from "@tallyho/tally-background/features"
 import SharedButton from "../../components/Shared/SharedButton"
 import SharedBackButton from "../../components/Shared/SharedBackButton"
 import OnboardingDerivationPathSelect from "../../components/Onboarding/OnboardingDerivationPathSelect"
