+++ ui/pages/Popup.tsx
@@ -12,7 +12,7 @@ import { Store } from "webext-redux"
 import { Provider } from "react-redux"
 import { TransitionGroup, CSSTransition } from "react-transition-group"
 import { isAllowedQueryParamPage } from "@tallyho/provider-bridge-shared"
-import { PERSIST_UI_LOCATION } from "@tallyho/tally-background/features/features"
+import { PERSIST_UI_LOCATION } from "@tallyho/tally-background/features"
 import { runtime } from "webextension-polyfill"
 import { popupMonitorPortName } from "@tallyho/tally-background/main"
 import { selectKeyringStatus } from "@tallyho/tally-background/redux-slices/selectors"
